# docs/swagger.yaml
basePath: /v1
definitions:
...
  internal.User:
    properties:
      created_at:
        description: CreatedAt
        type: string
      email:
        description: Email
        type: string
      id:
        description: ID
        type: integer
      is_activated:
        description: IsActivated
        type: boolean
      name:
        description: Name
        type: string
    type: object
  main.CreateAuthenticationTokenResponse:
    properties:
      token:
        type: string
    type: object
...
  main.CreatedUserResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/internal.User'
    type: object
...
  main.GetUserResponse:
    properties:
      user:
        $ref: '#/definitions/internal.User'
    type: object
...
  main.ResponseError:
    properties:
      error:
        description: Error
        type: string
    type: object
  main.ResponseMessage:
    properties:
      message:
        description: Message
        type: string
    type: object
...
  main.ViolationsMessage:
    properties:
      errors:
        additionalProperties:
          type: string
        description: errors
        type: object
    type: object
host: https://localhost:8080
info:
  contact:
    email: ahamerdev@gmail.com
    name: Ahmed Amer
  description: a simple movie reservation system api for educational purposes
  title: Movie Reservation System API
  version: "1.0"
paths:
...
  /tokens/activation:
    post:
      consumes:
      - application/json
      description: creates an activation token and sends it to the email
      parameters:
      - description: email of the user
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Creates an activation token
      tags:
      - tokens
    put:
      consumes:
      - application/json
      description: activates a user
      parameters:
      - description: token
        in: body
        name: token
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.User'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Activates a user
      tags:
      - tokens
  /tokens/authentication:
    post:
      consumes:
      - application/json
      description: creates an auth token
      parameters:
      - description: email
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: password
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.CreateAuthenticationTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Creates an auth token
      tags:
      - tokens
  /tokens/password-reset:
    post:
      consumes:
      - application/json
      description: creates a password-reset token and sends the token to the given
        email
      parameters:
      - description: email
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Creates a password-reset token
      tags:
      - tokens
    put:
      consumes:
      - application/json
      description: creates a password-reset token
      parameters:
      - description: email
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Creates a password-reset token
      tags:
      - tokens
  /users:
    post:
      consumes:
      - application/json
      description: creates a new user by name, email, password
      parameters:
      - description: name of the user
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: email of the user
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: password of the user
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.CreatedUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Creates a new user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: deletes the user
      parameters:
      - description: id of the user
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Delete User
      tags:
      - users
    get:
      consumes:
      - application/json
      description: gets The user Info by ID
      parameters:
      - description: id of the user
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.GetUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Get User Info
      tags:
      - users
    put:
      consumes:
      - application/json
      description: updates the user Info by ID
      parameters:
      - description: id of the user
        in: path
        name: id
        required: true
        type: integer
      - description: new name of the user
        in: body
        name: name
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Updates User Info
      tags:
      - users
swagger: "2.0"